requirements.txt
dj-database-url==0.4.2
Django==2.1.1
django-debug-toolbar==1.9.1
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
python-decouple==3.1
pytz==2018.5
sqlparse==0.2.4
